Fixes another major bug causing bogus cache behavior
When a state steals the cache from its descriptor, the operation can be
ignored when the cache is already referencing that state. However, the
code used a condition that the state's cache was pointing to the
descriptor's cache, which is clearly too weak; many states will
typically be looking at their descriptor's cache.